requirements.txt
flask
firebase-admin
pdfkit
jinja2
python-dotenv
requests
